[SCB-2008]remove archaius adapters
