--- foundations/foundation-config/src/main/java/org/apache/servicecomb/config/ConcurrentMapConfigurationExt.java
@@ -1,53 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.servicecomb.config;
-
-import java.util.Iterator;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-import org.apache.commons.configuration.AbstractConfiguration;
-
-import com.netflix.config.ConcurrentMapConfiguration;
-
-/**
- * Disable delimiter parsing for string
- */
-@SuppressWarnings("unchecked")
-public class ConcurrentMapConfigurationExt extends ConcurrentMapConfiguration {
-  public ConcurrentMapConfigurationExt() {
-    super();
-    setDelimiterParsingDisabled(true);
-  }
-
-  public ConcurrentMapConfigurationExt(Map<String, Object> mapToCopy) {
-    super();
-    setDelimiterParsingDisabled(true);
-    map = new ConcurrentHashMap<>(mapToCopy);
-  }
-
-  public ConcurrentMapConfigurationExt(AbstractConfiguration config) {
-    super();
-    config.setDelimiterParsingDisabled(true);
-    for (Iterator<String> i = config.getKeys(); i.hasNext(); ) {
-      String name = i.next();
-      Object value = config.getProperty(name);
-      map.put(name, value);
-    }
-  }
-}
