+++ foundations/foundation-config/src/main/java/org/apache/servicecomb/config/file/AbstractConfigLoader.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.config.archaius.sources;
+package org.apache.servicecomb.config.file;
 
 import java.io.IOException;
 import java.net.URL;
